improve reducer composition
