+++ src/containers/ChartViewContainer.tsx
@@ -3,7 +3,7 @@ import { Dispatch } from 'redux';
 
 import actions from '../actions';
 import ChartView from '../components/ChartView';
-import { Chart, StoreState } from '../store/types';
+import { Chart, StoreState } from '../shared/types';
 import { push } from 'react-router-redux';
 
 interface RouteProps {
