+++ src/containers/Root.tsx
@@ -1,25 +1,17 @@
 import * as React from 'react';
 import { Provider } from 'react-redux';
 import { Route } from 'react-router';
-import { applyMiddleware, createStore, combineReducers } from 'redux';
-import thunkMiddleware from 'redux-thunk';
 import createHistory from 'history/createBrowserHistory';
-import { ConnectedRouter, routerReducer, routerMiddleware } from 'react-router-redux';
+import { ConnectedRouter } from 'react-router-redux';
 
+import configureStore from '../store';
 import Layout from '../components/Layout';
-import reducer from '../reducers/index';
-import { StoreState } from '../store/types';
 import Dashboard from '../components/Dashboard';
 import ChartList from '../containers/ChartListContainer';
 import ChartView from '../containers/ChartViewContainer';
 
 const history = createHistory();
-const store = createStore<StoreState>(
-  combineReducers({
-    ...reducer,
-    router: routerReducer,
-  }),
-  applyMiddleware(thunkMiddleware, routerMiddleware(history)));
+const store = configureStore(history);
 
 class Root extends React.Component {
   render() {
